requirements.txt
Pillow==11.2.1
pystray==0.19.5
